Temporary workaround for email authentication bug
When an account is linked to multiple providers that includes an email provider, the user can no longer log in using email authentication. The bug seems to lie within the method, 'bestProviderFromProviderIDs', that finds the best provider to use for log in. The workaround performs a 'signIn' operation using the email provided by the user and a password that is guaranteed to fail, and then examining the return error. We want to know whether the email corresponds to a wrong password (meaning the user exists), or 'user not found'.